I tidied up the documentation a bit, mainly just spelling and idiomatic English.
There is only one change that needs comment: you sometimes used "wiki link" and
"wiki page" interchangeably. I think that the distinction between the two
things are clear, and it's not worthwhile remembering that "wiki link"
sometimes refers to the link, and sometimes to the page which is linked to.

However this usage is also in the command names, that use *Link throughout.
Maybe that's OK.